*Sigh*
Wow this blows... ok, basically fix both pidfiles and syslog.

The problem was that the previous code made a pidfile in /var/run/
that looked like this:  /var/run/dsdc_master[43432].pid

That wasn't useful for our rcscripts.

So to fix this, do something semi-gross, basically:

Use sfs::setprogname() before calling sfs::daemonnize() to set it
to just "sfs_master", then right after that, call sfs::setprogname()
again to set the progname to "sfs_master[pid]", this means that
/var/run will get "sfs_master.pid" and syslog will get "sfs_master[pid]:"