Fix a weird build issue.
Do not install dsdc_prot.x this was causing a weird effect where the
build of other projects that include "dsdc_prot.h" where re-compiling
the dsdc_prot.x instead of just sucking in the existing dsdc_prot.h.

I noticed this a while ago, but Jason poked me about it so it's fixed.